[Mass] Split MassTestCreation_test.cpp into individual test files for UniformMass and MeshMatrixMass
